+++ spec/integrations/pro/licensing/configure_with_valid_token_spec.rb
@@ -18,6 +18,7 @@ pro = Karafka::Pro
 assert_equal false, logs.include?('] ERROR -- : Your license expired')
 assert_equal false, logs.include?('Please reach us')
 assert Karafka.pro?
+assert const_visible?('Karafka::Pro::Processing::StrategySelector')
 assert const_visible?('Karafka::Pro::Processing::Coordinator')
 assert const_visible?('Karafka::Pro::Processing::Partitioner')
 assert const_visible?('Karafka::Pro::BaseConsumer')
@@ -34,6 +35,7 @@ assert const_visible?('Karafka::Pro::ActiveJob::Consumer')
 assert const_visible?('Karafka::Pro::ActiveJob::Dispatcher')
 assert const_visible?('Karafka::Pro::ActiveJob::JobOptionsContract')
 assert const_visible?('Karafka::Pro::PerformanceTracker')
+assert_equal pro::Processing::StrategySelector, config.processing.strategy_selector.class
 assert_equal pro::Processing::Partitioner, config.processing.partitioner_class
 assert_equal pro::Processing::Coordinator, config.processing.coordinator_class
 assert_equal pro::Processing::Scheduler, config.processing.scheduler.class